[MOV] fleet: Split model files according to coding guidelines
