+++ addons/fleet/models/fleet_vehicle.py
@@ -349,73 +349,3 @@ class FleetVehicle(models.Model):
             'domain': [('vehicle_id', '=', self.id)],
             'context': {'default_driver_id': self.driver_id.id, 'default_vehicle_id': self.id}
         }
-
-class FleetVehicleOdometer(models.Model):
-    _name = 'fleet.vehicle.odometer'
-    _description = 'Odometer log for a vehicle'
-    _order = 'date desc'
-
-    name = fields.Char(compute='_compute_vehicle_log_name', store=True)
-    date = fields.Date(default=fields.Date.context_today)
-    value = fields.Float('Odometer Value', group_operator="max")
-    vehicle_id = fields.Many2one('fleet.vehicle', 'Vehicle', required=True)
-    unit = fields.Selection(related='vehicle_id.odometer_unit', string="Unit", readonly=True)
-    driver_id = fields.Many2one(related="vehicle_id.driver_id", string="Driver", readonly=False)
-
-    @api.depends('vehicle_id', 'date')
-    def _compute_vehicle_log_name(self):
-        for record in self:
-            name = record.vehicle_id.name
-            if not name:
-                name = str(record.date)
-            elif record.date:
-                name += ' / ' + str(record.date)
-            record.name = name
-
-    @api.onchange('vehicle_id')
-    def _onchange_vehicle(self):
-        if self.vehicle_id:
-            self.unit = self.vehicle_id.odometer_unit
-
-
-class FleetVehicleState(models.Model):
-    _name = 'fleet.vehicle.state'
-    _order = 'sequence asc'
-    _description = 'Vehicle Status'
-
-    name = fields.Char(required=True, translate=True)
-    sequence = fields.Integer(help="Used to order the note stages")
-
-    _sql_constraints = [('fleet_state_name_unique', 'unique(name)', 'State name already exists')]
-
-
-class FleetVehicleTag(models.Model):
-    _name = 'fleet.vehicle.tag'
-    _description = 'Vehicle Tag'
-
-    name = fields.Char('Tag Name', required=True, translate=True)
-    color = fields.Integer('Color Index')
-
-    _sql_constraints = [('name_uniq', 'unique (name)', "Tag name already exists !")]
-
-
-class FleetServiceType(models.Model):
-    _name = 'fleet.service.type'
-    _description = 'Fleet Service Type'
-
-    name = fields.Char(required=True, translate=True)
-    category = fields.Selection([
-        ('contract', 'Contract'),
-        ('service', 'Service')
-        ], 'Category', required=True, help='Choose whether the service refer to contracts, vehicle services or both')
-
-
-class FleetVehicleAssignationLog(models.Model):
-    _name = "fleet.vehicle.assignation.log"
-    _description = "Drivers history on a vehicle"
-    _order = "create_date desc, date_start desc"
-
-    vehicle_id = fields.Many2one('fleet.vehicle', string="Vehicle", required=True)
-    driver_id = fields.Many2one('res.partner', string="Driver", required=True)
-    date_start = fields.Date(string="Start Date")
-    date_end = fields.Date(string="End Date")
