+++ addons/fleet/models/fleet_vehicle_log_contract.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo import api, fields, models, _
-from odoo.exceptions import UserError
-
 from dateutil.relativedelta import relativedelta
+
+from odoo import api, fields, models
+
 
 class FleetVehicleLogContract(models.Model):
     _inherit = ['mail.thread', 'mail.activity.mixin']
@@ -26,20 +26,22 @@ class FleetVehicleLogContract(models.Model):
     name = fields.Char(string='Name', compute='_compute_contract_name', store=True)
     active = fields.Boolean(default=True)
     user_id = fields.Many2one('res.users', 'Responsible', default=lambda self: self.env.user, index=True)
-    start_date = fields.Date('Contract Start Date', default=fields.Date.context_today,
+    start_date = fields.Date(
+        'Contract Start Date', default=fields.Date.context_today,
         help='Date when the coverage of the contract begins')
-    expiration_date = fields.Date('Contract Expiration Date', default=lambda self:
+    expiration_date = fields.Date(
+        'Contract Expiration Date', default=lambda self:
         self.compute_next_year_date(fields.Date.context_today(self)),
         help='Date when the coverage of the contract expirates (by default, one year after begin date)')
     days_left = fields.Integer(compute='_compute_days_left', string='Warning Date')
     insurer_id = fields.Many2one('res.partner', 'Vendor')
     purchaser_id = fields.Many2one(related='vehicle_id.driver_id', string='Current Driver')
     ins_ref = fields.Char('Reference', size=64, copy=False)
-    state = fields.Selection([
-        ('futur', 'Incoming'),
-        ('open', 'In Progress'),
-        ('expired', 'Expired'),
-        ('closed', 'Closed')
+    state = fields.Selection(
+        [('futur', 'Incoming'),
+         ('open', 'In Progress'),
+         ('expired', 'Expired'),
+         ('closed', 'Closed')
         ], 'Status', default='open', readonly=True,
         help='Choose whether the contract is still valid or not',
         tracking=True,
@@ -75,7 +77,7 @@ class FleetVehicleLogContract(models.Model):
                 today = fields.Date.from_string(fields.Date.today())
                 renew_date = fields.Date.from_string(record.expiration_date)
                 diff_time = (renew_date - today).days
-                record.days_left = diff_time > 0 and diff_time or 0
+                record.days_left = diff_time if diff_time > 0 else 0
             else:
                 record.days_left = -1
 
@@ -123,70 +125,3 @@ class FleetVehicleLogContract(models.Model):
 
     def run_scheduler(self):
         self.scheduler_manage_contract_expiration()
-
-class FleetVehicleLogServices(models.Model):
-    _name = 'fleet.vehicle.log.services'
-    _inherit = ['mail.thread', 'mail.activity.mixin']
-    _rec_name = 'service_type_id'
-    _description = 'Services for vehicles'
-
-    active = fields.Boolean(default=True)
-    vehicle_id = fields.Many2one('fleet.vehicle', 'Vehicle', required=True, help='Vehicle concerned by this log')
-    amount = fields.Monetary('Cost')
-    description = fields.Char('Description')
-    odometer_id = fields.Many2one('fleet.vehicle.odometer', 'Odometer', help='Odometer measure of the vehicle at the moment of this log')
-    odometer = fields.Float(compute="_get_odometer", inverse='_set_odometer', string='Odometer Value',
-        help='Odometer measure of the vehicle at the moment of this log')
-    odometer_unit = fields.Selection(related='vehicle_id.odometer_unit', string="Unit", readonly=True)
-    date = fields.Date(help='Date when the cost has been executed', default=fields.Date.context_today)
-    company_id = fields.Many2one('res.company', 'Company', default=lambda self: self.env.company)
-    currency_id = fields.Many2one('res.currency', related='company_id.currency_id')
-    purchaser_id = fields.Many2one('res.partner', string="Driver", compute='_compute_purchaser_id', readonly=False, store=True)
-    inv_ref = fields.Char('Vendor Reference')
-    vendor_id = fields.Many2one('res.partner', 'Vendor')
-    notes = fields.Text()
-    service_type_id = fields.Many2one(
-        'fleet.service.type', 'Service Type', required=True,
-        default=lambda self: self.env.ref('fleet.type_service_service_8', raise_if_not_found=False),
-    )
-    state = fields.Selection([
-        ('new', 'New'),
-        ('running', 'Running'),
-        ('done', 'Done'),
-        ('cancelled', 'Cancelled'),
-    ], default='new', string='Stage', group_expand='_expand_states')
-
-    def _get_odometer(self):
-        self.odometer = 0
-        for record in self:
-            if record.odometer_id:
-                record.odometer = record.odometer_id.value
-
-    def _set_odometer(self):
-        for record in self:
-            if not record.odometer:
-                raise UserError(_('Emptying the odometer value of a vehicle is not allowed.'))
-            odometer = self.env['fleet.vehicle.odometer'].create({
-                'value': record.odometer,
-                'date': record.date or fields.Date.context_today(record),
-                'vehicle_id': record.vehicle_id.id
-            })
-            self.odometer_id = odometer
-
-    @api.model_create_multi
-    def create(self, vals_list):
-        for data in vals_list:
-            if 'odometer' in data and not data['odometer']:
-                # if received value for odometer is 0, then remove it from the
-                # data as it would result to the creation of a
-                # odometer log with 0, which is to be avoided
-                del data['odometer']
-        return super(FleetVehicleLogServices, self).create(vals_list)
-
-    @api.depends('vehicle_id')
-    def _compute_purchaser_id(self):
-        for service in self:
-            service.purchaser_id = service.vehicle_id.driver_id
-
-    def _expand_states(self, states, domain, order):
-        return [key for key, dummy in type(self).state.selection]
