+++ addons/fleet/models/fleet_vehicle_model.py
@@ -48,43 +48,3 @@ class FleetVehicleModel(models.Model):
         }
 
         return view
-
-class FleetVehicleModelBrand(models.Model):
-    _name = 'fleet.vehicle.model.brand'
-    _description = 'Brand of the vehicle'
-    _order = 'name asc'
-
-    name = fields.Char('Make', required=True)
-    image_128 = fields.Image("Logo", max_width=128, max_height=128)
-    model_count = fields.Integer(compute="_compute_model_count", string="", store=True)
-    model_ids = fields.One2many('fleet.vehicle.model', 'brand_id')
-
-    @api.depends('model_ids')
-    def _compute_model_count(self):
-        Model = self.env['fleet.vehicle.model']
-        for record in self:
-            record.model_count = Model.search_count([('brand_id', '=', record.id)])
-
-    def action_brand_model(self):
-        self.ensure_one()
-        view = {
-            'type': 'ir.actions.act_window',
-            'view_mode': 'tree,form',
-            'res_model': 'fleet.vehicle.model',
-            'name': 'Models',
-            'context': {'search_default_brand_id': self.id, 'default_brand_id': self.id}
-        }
-
-        return view
-
-class FleetVehicleModelCategory(models.Model):
-    _name = 'fleet.vehicle.model.category'
-    _description = 'Category of the model'
-    _order = 'sequence asc, id asc'
-
-    _sql_constraints = [
-        ('name_uniq', 'UNIQUE (name)', 'Category name must be unique')
-    ]
-
-    name = fields.Char(required=True)
-    sequence = fields.Integer()
