+++ addons/fleet/models/fleet_vehicle_odometer.py
@@ -1,0 +1,32 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from odoo import api, fields, models
+
+
+class FleetVehicleOdometer(models.Model):
+    _name = 'fleet.vehicle.odometer'
+    _description = 'Odometer log for a vehicle'
+    _order = 'date desc'
+
+    name = fields.Char(compute='_compute_vehicle_log_name', store=True)
+    date = fields.Date(default=fields.Date.context_today)
+    value = fields.Float('Odometer Value', group_operator="max")
+    vehicle_id = fields.Many2one('fleet.vehicle', 'Vehicle', required=True)
+    unit = fields.Selection(related='vehicle_id.odometer_unit', string="Unit", readonly=True)
+    driver_id = fields.Many2one(related="vehicle_id.driver_id", string="Driver", readonly=False)
+
+    @api.depends('vehicle_id', 'date')
+    def _compute_vehicle_log_name(self):
+        for record in self:
+            name = record.vehicle_id.name
+            if not name:
+                name = str(record.date)
+            elif record.date:
+                name += ' / ' + str(record.date)
+            record.name = name
+
+    @api.onchange('vehicle_id')
+    def _onchange_vehicle(self):
+        if self.vehicle_id:
+            self.unit = self.vehicle_id.odometer_unit
