+++ addons/fleet/models/fleet_vehicle_state.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from odoo import fields, models
+
+
+class FleetVehicleState(models.Model):
+    _name = 'fleet.vehicle.state'
+    _order = 'sequence asc'
+    _description = 'Vehicle Status'
+
+    name = fields.Char(required=True, translate=True)
+    sequence = fields.Integer(help="Used to order the note stages")
+
+    _sql_constraints = [('fleet_state_name_unique', 'unique(name)', 'State name already exists')]
