+++ addons/fleet/models/fleet_vehicle_tag.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from odoo import fields, models
+
+
+class FleetVehicleTag(models.Model):
+    _name = 'fleet.vehicle.tag'
+    _description = 'Vehicle Tag'
+
+    name = fields.Char('Tag Name', required=True, translate=True)
+    color = fields.Integer('Color Index')
+
+    _sql_constraints = [('name_uniq', 'unique (name)', "Tag name already exists !")]
