+++ addons/fleet/models/res_config_settings.py
@@ -8,4 +8,3 @@ class ResConfigSettings(models.TransientModel):
     _inherit = ['res.config.settings']
 
     delay_alert_contract = fields.Integer(string='Delay alert contract outdated', default=30, config_parameter='hr_fleet.delay_alert_contract')
-
